Ajout classe DisplayValues pour réduire la complexité du code

+++ src/fr/univlille/info/J2/main/management/cells/Cell.java
@@ -37,7 +37,7 @@ public class Cell extends Rectangle{
 	 * @param gap_X La valeur de décalage en X pour la position de la cellule.
 	 * @param gap_Y La valeur de décalage en Y pour la position de la cellule.
 	 */
-	public Cell(int x, int y, int zoom, Color fill, int gap_X, int gap_Y) {
+	public Cell(int x, int y, double zoom, Color fill, double gap_X, double gap_Y) {
 		this(x,y,zoom,fill,fill, 0, gap_X, gap_Y);
 	}
 
@@ -54,7 +54,7 @@ public class Cell extends Rectangle{
 	 * @param gap_X La valeur de décalage en X pour la position de la cellule.
 	 * @param gap_Y La valeur de décalage en Y pour la position de la cellule.
 	 */
-	public Cell(int x, int y, int zoom, Color fill, Color stroke, int strokeWidth, int gap_X, int gap_Y) {
+	public Cell(int x, int y, double zoom, Color fill, Color stroke, int strokeWidth, double gap_X, double gap_Y) {
 		super(x*zoom+gap_X,y*zoom+gap_X,zoom, zoom);
 		this.coord=new Coordinate(y,x);
 		this.iv=new ImageView();
